app/src/test/java/com/cshawn/leetcodes/everyday/Q363Test.java
package com.cshawn.leetcodes.everyday;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;
import java.util.stream.Stream;

/**
 * @author C.Shawn
 * @date 2021/4/22 9:37 下午
 */
class Q363Test {

    public static class DataArgumentsProvider implements ArgumentsProvider {
        @Override
        public Stream<? extends Arguments> provideArguments(ExtensionContext context) throws Exception {
            return Stream.of(
                    Arguments.of(new int[][]{{1,0,1}, {0,-2,3}}, 2, 2),
                    Arguments.of(new int[][]{{2,2,-1}}, 3, 3)
            );
        }
    }

    @ParameterizedTest
    @ArgumentsSource(DataArgumentsProvider.class)
    void maxSumSubmatrix(int[][] matrix, int k, int result) {
        Assertions.assertEquals(result, new Q363().maxSumSubmatrix(matrix, k));
    }
}